Fix issue with Travis CI
It emerged today that the Travis CI build has broken since the last
merge to develop (at which point the build was working OK). It appears
that this is a result of the .travis.yaml file using "latest" for
the version of node_js.

Having done some research it seems the way to fix this is using the
latest LTS version of Node. At the time of writing this is 10.15.3.

See: https://docs.npmjs.com/try-the-latest-stable-version-of-node

I've therefore reverted to 10.x