Fix bug with duplicated ID's
A few months ago we encountered a bug which led to Mihai refactoring the application. This bug involved not letting us click on the search options on Discovery or TNA on the React component. It turns out that this was due to the duplication of IDs as React is rendering a desktop component and a mobile component. On TNA Base, the desktop component was loading first, and therefore the component seemed to work fine. On Discovery and TNA, the mobile component is loaded first, and therefore the desktop component broke, and we spent a good few hours trying to figure out why. I've made the ID's unique now by changing them based on their environment (i.e. website_search-desktop and website_search-mobile)